fix token story exports
